refactor(routes): Move routes into a separate file
As routes have five clear categories, splitting them into separate files
to group them by category makes locating them in the file system
easier. Related to issue #31
